linux: build FModSound, ADOImport, MiscDll, Input and Script
Five more modules now compile with GCC; DBFormat is the only one left before
Main/Game.

PlatformCompat gains the Win32 surface these modules need: CreateEvent/SetEvent/
ResetEvent (condition_variable-backed, sharing a polymorphic waitable base with
threads so WaitForSingleObject/CloseHandle still dispatch on a bare HANDLE),
LoadLibrary/FreeLibrary/GetProcAddress over dlopen, stricmp/strnicmp and
_wtof/_wtoi/_wtol. Misc links ${CMAKE_DL_LIBS} for the dlopen calls.

MSVC extensions the standard rejects:
 - BasicFactory.h / BasicDB.h: `name##::` pastes an identifier with `::`, which
   is not a valid preprocessing token; plain juxtaposition expands identically.
 - BasicDB.h: CDBPtr<T>::operator& is defined out-of-line after
   NDatabase::GetTable<>, whose name its body needs; Set/SetObject qualified
   with this-> for the dependent base.
 - DataConst.h: declare IsSuitableVariant before the template that calls it.
 - FMSoundStub.cpp: BASIC_REGISTER_CLASS moved outside namespace NFMSound,
   matching what FMSound.cpp already does.
 - lsaver.cpp: lua C function pointers are keyed by intptr_t, not int. Only the
   CLuaFuncID is ever serialized, so the save format is unchanged.
 - Input.h: forward-declare CDataStream instead of relying on include order.
 - Bind.cpp: its <dinput.h> include is Windows-only; the file itself never
   touches DirectInput.
 - Repo-wide: backslash include paths normalised to '/', and directory case in
   include paths corrected to match the on-disk names.

+++ Script/lsaver.cpp
@@ -2,8 +2,11 @@
 #include "lsaver.h"
 #include "lstring.h"
 //////////////////////////////////////////////////////////////////////////
-typedef unordered_map<CLuaFuncID, int> CLuaIDToFuncMap;
-typedef unordered_map<int, CLuaFuncID> CLuaFuncToIDMap;
+// The maps key C function pointers against their serialized CLuaFuncID; only
+// the ID ever reaches a save file, so widening the in-memory pointer key from
+// int to intptr_t (64-bit-safe) does not change the on-disk format.
+typedef unordered_map<CLuaFuncID, intptr_t> CLuaIDToFuncMap;
+typedef unordered_map<intptr_t, CLuaFuncID> CLuaFuncToIDMap;
 CLuaFuncToIDMap luaFuncToIDMap;
 CLuaIDToFuncMap luaIDToFuncMap;
 lua_State *pLUASaverState;
@@ -17,7 +20,7 @@ void lua_StartSerialize( lua_State *pL )
 //////////////////////////////////////////////////////////////////////////
 static void lua_GetID( CLuaFuncID *pID, lua_CFunction func )
 {
-	int nFunc = reinterpret_cast<int>( func );
+	intptr_t nFunc = reinterpret_cast<intptr_t>( func );
 	CLuaFuncToIDMap::iterator i = luaFuncToIDMap.find( nFunc );
 	ASSERT( i != luaFuncToIDMap.end() );  // unregistered lua C function!
 	*pID = i->second;
@@ -25,7 +28,7 @@ static void lua_GetID( CLuaFuncID *pID, lua_CFunction func )
 //////////////////////////////////////////////////////////////////////////
 static void lua_GetID( CLuaFuncID *pID, lua_Hook func )
 {
-	int nFunc = reinterpret_cast<int>( func );
+	intptr_t nFunc = reinterpret_cast<intptr_t>( func );
 	CLuaFuncToIDMap::iterator i = luaFuncToIDMap.find( nFunc );
 	ASSERT( i != luaFuncToIDMap.end() );  // unregistered lua C function!
 	*pID = i->second;
@@ -35,7 +38,7 @@ static lua_CFunction lua_GetFunc( const CLuaFuncID& id )
 { 
 	CLuaIDToFuncMap::iterator i = luaIDToFuncMap.find( id );
 	ASSERT( i != luaIDToFuncMap.end() );  // unregistered lua C function!
-	int nFunc = i->second;
+	intptr_t nFunc = i->second;
 	return ( lua_CFunction )nFunc;
 }
 //////////////////////////////////////////////////////////////////////////
@@ -43,20 +46,20 @@ static lua_Hook lua_GetHook( const CLuaFuncID& id )
 {
 	CLuaIDToFuncMap::iterator i = luaIDToFuncMap.find( id );
 	ASSERT( i != luaIDToFuncMap.end() );  // unregistered lua C function!
-	int nFunc = i->second;
+	intptr_t nFunc = i->second;
 	return ( lua_Hook )nFunc;
 }
 //////////////////////////////////////////////////////////////////////////
 void lua_RegisterFunc( lua_CFunction func, const CLuaFuncID& id )
 {
-	int nFunc = reinterpret_cast<int>( func );
+	intptr_t nFunc = reinterpret_cast<intptr_t>( func );
 	luaFuncToIDMap[ nFunc ] = id;
 	luaIDToFuncMap[ id ] = nFunc;
 }
 //////////////////////////////////////////////////////////////////////////
 void lua_RegisterFunc( lua_Hook func, const CLuaFuncID& id )
 {
-	int nFunc = reinterpret_cast<int>( func );
+	intptr_t nFunc = reinterpret_cast<intptr_t>( func );
 	luaFuncToIDMap[ nFunc ] = id;
 	luaIDToFuncMap[ id ] = nFunc;
 }
